+++ src/os/eeiop/cdvd/eecdvd.h
@@ -1,6 +1,8 @@
 #ifndef OS_EEIOP_CDVD_EECDVD_H
 #define OS_EEIOP_CDVD_EECDVD_H
 
+#include <stdint.h>
+
 #include "typedefs.h"
 
 typedef struct {
@@ -10,13 +12,13 @@ typedef struct {
 
 void CdvdInit();
 void CdvdInitSoftReset();
-int LoadReq(int file_no, u_int addr);
-u_int LoadReqGetAddr(int file_no, u_int addr, int *id);
+int LoadReq(int file_no, uint64_t addr);
+u_int LoadReqGetAddr(int file_no, uint64_t addr, int *id);
 int LoadReqSe(int file_no, u_char se_type);
-int LoadReqNSector(int file_no, int sector, int size, int addr);
-int LoadReqNFno(int file_no, int addr);
-int LoadReqBFno(int file_no, int addr);
-u_int LoadReqBFnoGetAddr(int file_no, int addr);
+int LoadReqNSector(int file_no, int sector, int size, int64_t addr);
+int LoadReqNFno(int file_no, int64_t addr);
+int LoadReqBFno(int file_no, int64_t addr);
+u_int LoadReqBFnoGetAddr(int file_no, int64_t addr);
 int IsLoadEndAll();
 int IsLoadEnd(int id);
 void LoadEndFlgRenew();
